added _isActive to the class CourseEditionEnrollment; added isActive to the CourseEditionEnrollment constructor; added the method isEnrollmentActive(); added the method deactivateEnrollment(); change the method in US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller removeStudentEnrolment changed the test of Us28

+++ src/main/java/PAI/controller/US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller.java
@@ -13,12 +13,9 @@ public class US28_RemoveTheEnrollmentOfAStudentInACourseEdition_Controller {
         this.courseEditionEnrollmentRepository = courseEditionEnrollmentRepository;
     }
     public boolean removeStudentEnrolment(Student student, CourseEdition courseEdition) {
-        try {
-            courseEditionEnrollmentRepository.removeEnrollment(student, courseEdition);
-
-        } catch (Exception e) {
-            return false;
+        if(student==null || courseEdition==null){
+            throw new IllegalArgumentException("Student and CourseEdition cannot be null");
         }
-        return true;
+           return courseEditionEnrollmentRepository.removeEnrollment(student, courseEdition);
     }
 }
